finn-t/requirements-working.txt
annotated-types==0.7.0
attrs==26.1.0
bitarray==3.8.1
bitstring==4.3.1
brevitas==0.12.1
cffi==2.0.0
click==8.3.2
clize==5.0.2
cmake==4.0.3
coloredlogs==15.0.1
dependencies==2.0.1
docutils==0.22.4
einops==0.8.2
filelock==3.25.2
finn-plus==1.4.0
flatbuffers==25.12.19
fsspec==2026.3.0
h5py==3.16.0
humanfriendly==10.0
importlib_metadata==9.0.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mashumaro==3.20
mdurl==0.1.2
mip==1.13.0
ml_dtypes==0.5.4
mpmath==1.3.0
netron==9.0.1
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
od==2.0.2
onnx==1.17.0
onnx-ir==0.1.12
-e git+https://github.com/iksnagreb/onnx-passes.git@e5ff48c2292f83d490b0aed11437d953193701d6#egg=onnx_passes
onnxoptimizer==0.3.13
onnxruntime==1.20.1
onnxscript==0.5.4
packaging==26.0
protobuf==4.25.8
pybind11==3.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
PyYAML==6.0.3
qonnx==1.0.0
rich==13.9.4
scipy==1.15.3
sigtools==4.0.1
sympy==1.14.0
termcolor==2.5.0
toposort==1.10
torch==2.3.1
tqdm==4.67.3
triton==2.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
unfoldNd==0.2.3
zipp==3.23.0
